requirements.txt
PyJWT==1.7.1
alembic==1.0.10
aniso8601==7.0.0
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Limiter==1.0.1
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.1
limits==1.3
Mako==1.0.12
MarkupSafe==1.1.1
marshmallow==2.19.5
psycopg2-binary==2.8.3
pycparser==2.19
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.5
webargs==5.3.2
Werkzeug==0.15.4
